Use one client/connection per network/server
Prior to this change a new client and connection was created for every
channel/room even if those channels/rooms happened to be on the same
network/server.

With this change only one client and connection is created per server.
The configuration format has been changed to specify the multiple
channels/rooms for a network/server as a list under the entry for a
network/server.